Break up the fileStore file into multiple files
